adapt the dh distribution the new signal handling

+++ launch_dh.sh
@@ -46,43 +46,35 @@ clush -w @client -b ./prototype_MDS/client_launch.sh $(($nb_srv)) /media/traces/
 printf "setup finished\n"
 
 current_step=0
+turn=0
+
 while [[ $current_step -lt $total_step ]]
 do
-    for((turn = 0; turn < rebalance; turn++))
-    do
-        #launch a step of traces
-        clush -w @client -b ./prototype_MDS/client_launch.sh $(($nb_srv)) $traces_path/changelog-$current_step /mnt/dh
-        ((current_step++))
+    #launch a step of traces
+    clush -w @client -b ./prototype_MDS/client_launch.sh $(($nb_srv)) $traces_path/changelog-$current_step /mnt/dh
+    ((current_step++))
 
+    if [ $turn eq $rebalance ]
+    then
+        #launch redistribution
+        clush -w @srv 'kill -s SIGUSR2 `/usr/sbin/pidof ./prototype_MDS/bin/server`'
+        clush -w vm0 'kill -s SIGUSR2 `/usr/sbin/pidof ./prototype_MDS/bin/manager`'
+        turn=0
+    else
         #mesuring
         clush -w @srv 'kill -s SIGUSR1 `/usr/sbin/pidof ./prototype_MDS/bin/server`'
+    fi
 
-        #wait for file creation "vm<id_srv>USR1"
-        for ((i = 0; i < nb_srv; i++))
-        do
-            if ! [ -f "/media/tmp_ack/dh/vm$(($i))USR1" ]
-            then
-                ((i--))
-                sleep 1
-            fi
-        done
-        rm /media/tmp_ack/dh/*USR1
-
-        printf "step $current_step finished\n"
-    done
-
-    #launch redistribution
-    clush -w @srv 'kill -s SIGUSR2 `/usr/sbin/pidof ./prototype_MDS/bin/server`'
-    clush -w vm0 'kill -s SIGUSR2 `/usr/sbin/pidof ./prototype_MDS/bin/manager`'
-
-    #wait for file creation "id_svrUSR2"
+    #wait for file creation "vm<id_srv>USR"
     for ((i = 0; i < nb_srv; i++))
     do
-        if ! [ -f "/media/tmp_ack/dh/vm$(($i))USR2" ]
+        if ! [ -f "/media/tmp_ack/dh/vm$(($i))USR" ]
         then
             ((i--))
             sleep 1
         fi
-    done
-    rm /media/tmp_ack/dh/*USR2
+        done
+    rm /media/tmp_ack/dh/*USR
+
+    printf "step $current_step finished\n"
 done
